New React Site
This is the react base, It is not aesthetically pleasing...YET.  I'm having to refresh my memory on how all these bits and peices interact. I am watching a tutorial and coding along, but I intend on changing A LOT of things by the end of the next sprint (Thursday).