Fixed Bugs
Fixed the Context bugs, this part of the project was actually quite difficult. All the links are working, and I have a working filter, Sorry Kristin for making it seem like you're on sale. haha. This will be a feature to go through all the tpes of accounts. I still am working on a FUNCTIONAL login page. and registration. I'm going to have to see how to link all the components together to show the user information etc. Thanks for the patience guys, this is a lot more strenuous than just building a site from scratch and getting it hosted. Appreciate your encouragement.