Services.css
Services CSS file created and made to fit in smaller screen, let me know if it works nicely on your end, then I can alter some of the very specific definitions I placed. I'm looking for ways to make it more generic and not so specific to my browser response and resolution.