Something frustrating
each browser reacts differently to resolution? example: On a smaller screen, Mozilla will format the About us page perfectly, however on Chrome, the divs are separated differently. Im not sure how to explain it, but it is incredibly annoying to fix something in one screen and it not render properl y in another browser. :)